+++ BaseUriTest.php
@@ -357,6 +357,10 @@ final class BaseUriTest extends TestCase
                 'uri' => 'https://[1050:0000:0000:0000:0005:0000:300c:326b]:443/',
                 'expectedOrigin' => 'https://[1050::5:0:300c:326b]',
             ],
+            'normalized ipv4 non decimal notation' => [
+                'uri' => 'https://0xc0a821/foo/bar/',
+                'expectedOrigin' => 'https://0.192.168.33',
+            ],
         ];
     }